test(evals): merge den-sidebar-footer-pin into den-sidebar-ia
