+++ evals/specs/den-sidebar-ia.e2e.test.ts
@@ -21,6 +21,18 @@ type SidebarSection = {
   children: string[];
 };
 
+interface SidebarPinLayout {
+  viewportHeight: number;
+  navScrollHeight: number;
+  navClientHeight: number;
+  navOverflowY: string;
+  asideScrollTop: number;
+  asideClientHeight: number;
+  asideScrollHeight: number;
+  footerTop: number;
+  footerBottom: number;
+}
+
 function isRecord(value: unknown): value is Record<string, unknown> {
   return typeof value === "object" && value !== null && !Array.isArray(value);
 }
@@ -42,6 +54,36 @@ function parseSections(value: unknown): SidebarSection[] {
     }
     return { name: entry.name, items: entry.items, children: entry.children };
   });
+}
+
+function parsePinLayout(value: unknown): SidebarPinLayout {
+  if (!isRecord(value)) {
+    throw new Error(`Sidebar pin layout was not found: ${JSON.stringify(value)}`);
+  }
+  if (
+    typeof value.viewportHeight !== "number"
+    || typeof value.navScrollHeight !== "number"
+    || typeof value.navClientHeight !== "number"
+    || typeof value.navOverflowY !== "string"
+    || typeof value.asideScrollTop !== "number"
+    || typeof value.asideClientHeight !== "number"
+    || typeof value.asideScrollHeight !== "number"
+    || typeof value.footerTop !== "number"
+    || typeof value.footerBottom !== "number"
+  ) {
+    throw new Error(`Sidebar pin layout had an unexpected shape: ${JSON.stringify(value)}`);
+  }
+  return {
+    viewportHeight: value.viewportHeight,
+    navScrollHeight: value.navScrollHeight,
+    navClientHeight: value.navClientHeight,
+    navOverflowY: value.navOverflowY,
+    asideScrollTop: value.asideScrollTop,
+    asideClientHeight: value.asideClientHeight,
+    asideScrollHeight: value.asideScrollHeight,
+    footerTop: value.footerTop,
+    footerBottom: value.footerBottom,
+  };
 }
 
 function itemNames(sections: SidebarSection[]): string[] {
@@ -74,6 +116,27 @@ const readSidebar = `(() => {
       (link.textContent ?? "").replace(/\\s+/g, " ").trim()
     ),
   }));
+})()`;
+
+const readPinLayout = `(() => {
+  const nav = document.querySelector('[data-testid="den-org-sidebar"]');
+  const footer = document.querySelector('[data-testid="den-org-sidebar-footer"]');
+  const aside = nav?.closest("aside");
+  if (!(nav instanceof HTMLElement) || !(footer instanceof HTMLElement) || !(aside instanceof HTMLElement)) {
+    return null;
+  }
+  const footerRect = footer.getBoundingClientRect();
+  return {
+    viewportHeight: window.innerHeight,
+    navScrollHeight: nav.scrollHeight,
+    navClientHeight: nav.clientHeight,
+    navOverflowY: getComputedStyle(nav).overflowY,
+    asideScrollTop: aside.scrollTop,
+    asideClientHeight: aside.clientHeight,
+    asideScrollHeight: aside.scrollHeight,
+    footerTop: footerRect.top,
+    footerBottom: footerRect.bottom,
+  };
 })()`;
 
 test(title, async ({ evidence, place }) => {
@@ -182,6 +245,33 @@ test(title, async ({ evidence, place }) => {
     ]);
     expect(seen.ok, seen.why).toBe(true);
   }
+
+  // The same signed-in admin session must keep the workspace switcher pinned
+  // when the expanded Settings tree is taller than the viewport.
+  await browser.client.send("Emulation.setDeviceMetricsOverride", {
+    width: 1440,
+    height: 720,
+    deviceScaleFactor: 1,
+    mobile: false,
+  });
+  await waitFor(browser, `window.innerHeight === 720`, {
+    timeoutMs: 30_000,
+    label: "admin sidebar reflowed to the short viewport",
+  });
+  const pinLayout = parsePinLayout(await evalIn(browser, readPinLayout));
+  const footerInView = pinLayout.footerTop >= 0 && pinLayout.footerBottom <= pinLayout.viewportHeight;
+  const navIsScroller = pinLayout.navScrollHeight > pinLayout.navClientHeight + 8
+    && (pinLayout.navOverflowY === "auto" || pinLayout.navOverflowY === "scroll");
+  const asideDoesNotOverflow = pinLayout.asideScrollTop === 0
+    && pinLayout.asideScrollHeight <= pinLayout.asideClientHeight + 1;
+  expect(footerInView).toBe(true);
+  expect(navIsScroller).toBe(true);
+  expect(asideDoesNotOverflow).toBe(true);
+  evidence.recordAssertionEvidence(
+    "At 720px, the workspace switcher stays visible while only the sidebar navigation scrolls",
+    JSON.stringify(pinLayout),
+    footerInView && navIsScroller && asideDoesNotOverflow,
+  );
 
   const memberTokenStored = await evalIn(browser, `(() => {
     localStorage.setItem("openwork:web:auth-token", ${JSON.stringify(member.token)});
